On .NET 6 (Seq 2022.1), fall back to using NLS/avoid IANA time zone names in test
